// src/app/core/model/album.ts
// Generated by https://quicktype.io


export interface AlbumsSearchResponse {
  albums: PagingObject<Album>;
}

export interface AlbumsResponse {
  albums: Album[];
}

export interface Album {
  album_type:             string;
  total_tracks:           number;
  // available_markets:      string[];
  // external_urls:          ExternalUrls;
  href:                   string;
  id:                     string;
  images:                 Image[];
  name:                   string;
  release_date:           string;
  release_date_precision: string;
  // restrictions:           Restrictions;
  type:                   "album";
  uri:                    string;
  artists:                Artist[];
  tracks?:                PagingObject<Track>;
}

export interface AlbumItemView {
  id: string;
  images: Image[];
  name: string;
}

export interface Artist {
  external_urls: ExternalUrls;
  followers:     Followers;
  genres:        string[];
  href:          string;
  id:            string;
  images:        Image[];
  name:          string;
  popularity:    number;
  type:          "artist";
  uri:           string;
}

export interface ExternalUrls {
  spotify: string;
}

export interface Followers {
  href:  string;
  total: number;
}

export interface Image {
  url:    string;
  height: number;
  width:  number;
}

export interface Restrictions {
  reason: string;
}

export interface PagingObject<T> {
  href:     string;
  items:    T[];
  limit:    number;
  next:     string;
  offset:   number;
  previous: string;
  total:    number;
}

export interface Track {
  id:            string;
  release_date:  string;
  name:          string;
  duration_ms:   string;
  type:          "artist";
}